backend/requirements.txt
# ─── Core ───
torch>=2.0.0
transformers>=4.38.0
sentence-transformers>=2.5.0

# ─── PDF Processing ───
pymupdf>=1.23.0

# ─── Vector DB ───
chromadb>=0.4.22

# ─── Retrieval ───
numpy>=1.24.0
requests>=2.31.0

# ─── API Server (Production) ───
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# ─── Demo UI ───
streamlit>=1.30.0

# ─── Evaluation (optional) ───
# ragas>=0.1.0  # pip install ragas

# ─── Auth + DB ───
python-docx>=0.8.11
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
psycopg2-binary>=2.9.0
